requirements.txt
# Core Document Processing
pdfplumber>=0.9.0           # PDF text extraction with layout preservation
python-docx>=1.0.0          # Microsoft Word document processing
easyocr>=1.6.0              # Optical Character Recognition for images
Pillow>=9.0.0               # Image processing support

# LLM Integration
openai>=1.0.0               # OpenAI API client
requests>=2.25.0            # HTTP requests for Ollama API

# Configuration and Data
PyYAML>=6.0                 # YAML configuration parsing
typing-extensions>=4.0.0    # Type hints compatibility

# Optional Dependencies (uncomment if needed)
# nltk>=3.8                 # Enhanced text processing
# httpx>=0.24.0             # Alternative HTTP client